Revise review comments
